==> scripts/s01_read_data_feb2016.Rmd <==
---
title: "Read data, wecare feb2016"
author: "Alexey Larionov"
output: html_document
params:
  source_data: ""
  interim_data: ""
---

started: 01Mar2016  
last updated: 04Aug2016

# start_section

```{r start_section}

# Time stamp
Sys.time()

# Do NOT clean-up
#rm(list=ls()) # this would remove params object!
graphics.off()

# Read parameters
source_data_folder <- params$source_data
interim_data_folder <- params$interim_data

print(source_data_folder)
print(interim_data_folder)

# Required libraries
library(ggplot2)

# For debugging
#source_data_folder <- "/scratch/medgen/users/alexey/wecare_aug2016/source_data"
#interim_data_folder <- "/scratch/medgen/users/alexey/wecare_aug2016/interim_data"
#setwd("/scratch/medgen/scripts/wecare_skat_08.16/scripts")

```

# read_data

```{r read_data}

# Compile file names
covar_file <- paste(source_data_folder, "covar.txt", sep="/")
demographics_file <- paste(source_data_folder, "WECARE.Exome.DemographicVariables.txt", sep="/")
samples_file <- paste(source_data_folder, "samples_ids.txt", sep="/")
vv_file <- paste(source_data_folder, "wecare_feb2016_std_filter_VV_biallelic.txt", sep="/")
gt_file <- paste(source_data_folder, "wecare_feb2016_std_filter_GT_biallelic_add.txt", sep="/")
gq_file <- paste(source_data_folder, "wecare_feb2016_std_filter_GQ_biallelic.txt", sep="/")
dp_file <- paste(source_data_folder, "wecare_feb2016_std_filter_DP_biallelic.txt", sep="/")
ad_file <- paste(source_data_folder, "wecare_feb2016_std_filter_AD_biallelic.txt", sep="/")

# Read 
covar.df <- read.table(covar_file, sep="\t", header=TRUE, quote="")
demographics.df <- read.table(demographics_file, sep="\t", header=TRUE, quote="")
samples.df <- read.table(samples_file, sep="\t", header=TRUE, quote="")
vv.df <- read.table(vv_file, header=TRUE, row.names=1, sep="\t", quote="")
gt.df <- read.table(gt_file, header=TRUE, row.names=1, sep="\t", quote="")
gq.df <- read.table(gq_file, header=TRUE, row.names=1, sep="\t", quote="")
dp.df <- read.table(dp_file, header=TRUE, row.names=1, sep="\t", quote="")
ad.df <- read.table(ad_file, header=TRUE, row.names=1, sep="\t", quote="")

rm(source_data_folder, covar_file, demographics_file, samples_file, vv_file, gt_file, gq_file, dp_file, ad_file)

```

# check_data

```{r check_data}

dim(covar.df)
str(covar.df)
covar.df[1:5,1:5]

dim(demographics.df)
str(demographics.df)
demographics.df[1:5,1:5]

dim(samples.df)
str(samples.df)
samples.df[1:5,]

dim(vv.df)
str(vv.df)
vv.df[1:5,1:5]

dim(gt.df)
str(gt.df, list.len=5)
gt.df[1:5,1:5]

dim(gq.df)
str(gq.df, list.len=5)
gq.df[1:5,1:5]

dim(dp.df)
str(dp.df, list.len=5)
dp.df[1:5,1:5]

dim(ad.df)
str(ad.df, list.len=5)
ad.df[1:15,1:5]

```

# explore_age
Inclusion criteria: age at 1st Ds < 55

```{r explore_age}

ggplot(covar.df, aes(sub_dx_age)) +
  geom_histogram(binwidth = 5, colour="black", alpha=0.25) + 
  aes(fill=as.factor(cc)) +
  ggtitle("Age at the 1st diagnosis in cases and controls")

```

# explore_time_to_2nd_ca
Inclusion criteria: 2nd Ds should be 1yr+ later than the 1st Ds
There are two fields that look appropriate ...

```{r explore_time_to_2nd_ca}

ggplot(covar.df, aes(offset)) +
  geom_histogram(breaks=c(0:8), right=FALSE, colour="black", alpha=0.25) + 
  aes(fill=as.factor(cc)) +
  ggtitle("covar: offset")

ggplot(demographics.df[c(-59, -502,-503,-504,-505),], aes(rstime)) +
  geom_histogram(breaks=c(0:8), right=FALSE, colour="black", alpha=0.25) + 
  aes(fill=as.factor(cc.y)) +
  ggtitle("demographics: rstime")

```

# split_ad_table

```{r split_ad_table}

# Convert source data frame to matrix
ad.mx <- as.matrix(ad.df)
rm(ad.df)

# Prepare matrices for output
ref.mx <- ad.mx
alt.mx <- ad.mx

# Split ad into two tables (long step)
for (row in 1:nrow(ad.mx)){
  for (col in 1:ncol(ad.mx)){
    ref.alt <- strsplit(ad.mx[row,col],",")[[1]]
    ref.mx[row,col] <- ref.alt[1]
    alt.mx[row,col] <- ref.alt[2]
  }
}

# Clean-up
rm(row, col, ref.alt)

# Explore results
dim(ad.mx)
dim(ref.mx)
dim(alt.mx)

ad.mx[1:15,1:5]
ref.mx[1:15,1:5]
alt.mx[1:15,1:5]

# Clean-up
rm(ad.mx)

```

# save_data

```{r save_data}

interim_data_file <- paste(interim_data_folder, "s01_read_data_feb2016.RData", sep="/")
save.image(file=interim_data_file)

```

# final_section

```{r final_section}

ls()
sessionInfo()
Sys.time()

```
