==> scripts/s02_filter_variants_feb2016.Rmd <==
---
title: "Apply variants filters, wecare feb2016"
author: "Alexey Larionov"
output: html_document
params:
  interim_data: ""
---

started: 01Mar2016  
last updated: 02Aug2016

# start_section

```{r start_section}

# Time stamp
Sys.time()

# Do NOT clean-up at this stage
#rm(list=ls()) # this would remove params object!
graphics.off()

# Read parameters
interim_data_folder <- params$interim_data

# Remove params object to avoid interference with load() function
rm(params)

# For debugging
#interim_data_folder <- "/scratch/medgen/users/alexey/wecare_aug2016/interim_data"
#setwd("/scratch/medgen/scripts/wecare_skat_08.16/scripts")

```

# load_and_check_data

```{r load_and_check_data}

interim_data_file <- paste(interim_data_folder, "s01_read_data_feb2016.RData", sep="/")
load(file=interim_data_file)
ls()

dim(covar.df)
str(covar.df)
covar.df[1:5,1:5]

dim(samples.df)
str(samples.df)
samples.df[1:5,]

dim(vv.df)
str(vv.df)
vv.df[1:5,1:5]

dim(gt.df)
str(gt.df, list.len=5)
gt.df[1:5,1:5]

dim(gq.df)
str(gq.df, list.len=5)
gq.df[1:5,1:5]

dim(dp.df)
str(dp.df, list.len=5)
dp.df[1:5,1:5]

dim(ref.mx)
class(ref.mx)
ref.mx[1:15,1:5]

dim(alt.mx)
class(alt.mx)
alt.mx[1:15,1:5]

# Check consistence of rownames
sum(rownames(gt.df) != rownames(vv.df), na.rm=TRUE)

sum(rownames(gt.df) != rownames(gq.df), na.rm=TRUE)
sum(rownames(gt.df) != rownames(dp.df), na.rm=TRUE)
sum(rownames(gt.df) != rownames(ref.mx), na.rm=TRUE)
sum(rownames(gt.df) != rownames(alt.mx), na.rm=TRUE)

# Check consistence of colnames
sum(substr(colnames(gt.df),0,6) != substr(colnames(gq.df),0,6), na.rm=TRUE)
sum(substr(colnames(gt.df),0,6) != substr(colnames(dp.df),0,6), na.rm=TRUE)
sum(substr(colnames(gt.df),0,6) != substr(colnames(ref.mx),0,6), na.rm=TRUE)
sum(substr(colnames(gt.df),0,6) != substr(colnames(alt.mx),0,6), na.rm=TRUE)

```

# filter_variants

```{r filter_variants}

# Genoptypes NA rate before variants filtering
sum(is.na(gt.df))/(dim(gt.df)[1]*dim(gt.df)[2]) # ~6%

vars.pf <- vv.df$FILTER == "PASS"
sum(vars.pf)

vv.pf.df <- vv.df[vars.pf,]
dim(vv.pf.df)
rm(vv.df)

gt.pf.df <- gt.df[vars.pf,]
dim(gt.pf.df)
rm(gt.df)

gq.pf.df <- gq.df[vars.pf,]
dim(gq.pf.df)
rm(gq.df)

dp.pf.df <- dp.df[vars.pf,]
dim(dp.pf.df)
rm(dp.df)

ref.pf.mx <- ref.mx[vars.pf,]
dim(ref.pf.mx)
rm(ref.mx)

alt.pf.mx <- alt.mx[vars.pf,]
dim(alt.pf.mx)
rm(alt.mx)

rm(vars.pf)

# Genoptypes NA rate after variants filtering
sum(is.na(gt.pf.df))/(dim(gt.pf.df)[1]*dim(gt.pf.df)[2]) # <1%

```

# save_data

```{r save_data}

interim_data_file <- paste(interim_data_folder, "s02_filter_variants_feb2016.RData", sep="/")
save.image(file=interim_data_file)

```

# final_section

```{r final_section}

ls()
sessionInfo()
Sys.time()

```
